Rebar3 include location
rebar3 moves the build around a bit, and all the build related stuff is
now stored in `_build` including the fetched dependencies. Currently
this means that when using templates like for example for webmachine,
syntaxerl will not find the `hrl` files. Adding the path rebar3 works
with to the defaults fixes this and this work again. I think do to the
naming of `_build/default/lib` getting a false positive in pre rebar3
project is low enough to just add this as this.